requirements.txt
python-decouple
requests
tweepy